fix, TechSelectBlock 컴포넌트 확장성 수정
